Pridanie: Implementácia triedy GameUtils pre spracovanie herných funkcií a manipuláciu s oknom. Zmena: Presun funkcie _resizeWindow do súboru functions.js a jej exportovanie. Odstránenie: Odstránenie nepoužívaného kódu v súbore gameUtils.js.
